requirements.txt
fastapi==0.122.0
uvicorn==0.38.0
redis==7.1.0
psycopg2-binary==2.9.11
alembic==1.17.2
